pyproject.toml
[project]
name = "ai_gemini_example"
version = "0.1.0"
description = "A python example for Google Gemini AI"
authors = [
    {name = "Ryan Liu", email = "azhai@126.com"},
]
dependencies = [
    "google-generativeai>=0.7.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
